LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointer_boundary_events_after_removing_last_over_element-expected.txt


FAIL After removing the last over element, redundant pointerenter events should not be fired on the ancestors assert_equals: expected "pointerleave@div#parent, pointerleave@div#grandparent, pointerover@body, pointermove@body" but got "pointerout@div#child, pointerleave@div#child, pointerover@body, pointerenter@body, pointermove@body"
FAIL After removing the root element in the shadow under the cursor, pointerleave events should be targeted outside the shadow, but redundant pointerenter events should not be fired assert_equals: expected "pointerleave@div#shadowHost, pointerleave@div#containerOfShadowHost, pointerover@body, pointermove@body" but got "pointerout@div#divInShadow, pointerleave@div#divInShadow, pointerover@body, pointerenter@body, pointermove@body"
